Implemented Event and EventDispatcher with GUI controls
Removed all previous GUI classes and redesigned the system based on
Lythom's HaxePunk GUI. The previous system worked by colliding a mouse
pointer Entity with GUI controls to imitate events. The new system
implements flash.events.Event and flash.events.EventDispatcher to do it
the "right" way.

Control is the parent class to every GUI control element, such as the
Button class. Control extends Entity in order to be usable on Scenes,
and implements IEventDispatcher so that Events can be dispatched and
listened for.

ControlEvent is a custom Event with the added variable of type. Adapted
directly from Lythom. May be dropped in the future.